MSVC: replaced uint with uint32_t, and fixed other compilation errors.
More errors remains, but I'll handle them later.

+++ src/forms/AudioTestForm.cpp
@@ -95,7 +95,7 @@ void ebox::AudioTestForm::drawAudioPanel()
 void ebox::AudioTestForm::drawAudioInfo()
 {
     ImGui::BeginChild("test_audio_2", {-1, 250}, true, 0);
-    ImGui::BeginChild("test_audio_2_sub1", {(getCurrentWindowSize().x / 2), -1}, true, 0);
+    ImGui::BeginChild("test_audio_2_sub1", {(static_cast<float>(getCurrentWindowSize().x) / 2), -1}, true, 0);
     ImGui::Text(fmt::format("Track: {0} of {1}", m_stream.getInfo().getTrackNumber() + 1, m_stream.getInfo().getNumberOfTracks()).c_str());
     ImGui::Text(fmt::format("Song: {0}", m_stream.getInfo().getSong()).c_str());
     ImGui::Text(fmt::format("Intro length: {0}", m_stream.getInfo().getIntroLength()).c_str());
